requirements.txt
# Core
torch>=2.4.0
transformers>=4.45.0
datasets>=3.0.0
accelerate>=0.34.0
peft>=0.13.0
trl>=0.12.0

# GRPO training
verl>=0.7.0
ray>=2.40.0
hydra-core>=1.3.0

# Inference
vllm>=0.6.0
openai>=1.50.0

# Reward model serving
openrlhf>=0.5.0
aiohttp>=3.10.0

# Data processing
pyarrow>=17.0.0

# Evaluation and analysis
nltk>=3.9.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.14.0
scikit-learn>=1.5.0
matplotlib>=3.9.0
bert-score>=0.3.13

# Utilities
absl-py>=2.1.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
termcolor>=2.4.0
tqdm>=4.66.0
jsonlines>=4.0.0
